scripts/feature-extraction/requirements.txt
cython
h5py==3.1.*
imgaug==0.3.*
matplotlib==3.1.*
numpy==1.22.*
opencv-python-headless==4.5.4.*
Pillow==9.0.*
imageio==2.10.*
scikit-image==0.16.*
scipy==1.1.*
tensorflow==2.6.4
